video player page separation
