+++ lib/screen/home_page.dart
@@ -1,8 +1,7 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:flutter/material.dart';
-import 'package:video_player/video_player.dart';
-
+import 'package:canlitvizle/screen/video_player_page.dart';
 import '../model/channel.dart';
 
 class HomePage extends StatefulWidget {
@@ -79,51 +78,4 @@ class _HomePageState extends State<HomePage> {
   }
 }
 
-class VideoPlayerPage extends StatefulWidget {
-  final String title;
-  final String streamUrl;
-  final String imgUrl;
 
-  VideoPlayerPage(this.title, this.streamUrl, this.imgUrl);
-
-  @override
-  _VideoPlayerPageState createState() => _VideoPlayerPageState();
-}
-
-class _VideoPlayerPageState extends State<VideoPlayerPage> {
-  late VideoPlayerController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = VideoPlayerController.network(widget.streamUrl)
-      ..initialize().then((_) {
-        setState(() {
-          _controller.play();
-        });
-      });
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _controller.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: _controller.value.isInitialized
-            ? AspectRatio(
-                aspectRatio: _controller.value.aspectRatio,
-                child: VideoPlayer(_controller),
-              )
-            : CircularProgressIndicator(),
-      ),
-    );
-  }
-}
